changed vm scripts

+++ home/.scripts/vm_shutdown.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
-nc 127.0.0.1 1234 << END
+
+if [[ -z $(netstat -tln | grep 127.0.0.1:1234) ]]; then
+  echo "VM not running"
+  exit
+fi
+
+killall synergy synergyc
+
+nc -N 127.0.0.1 1234 << END
 device_del usb_kbd
 device_del usb_mouse
 system_powerdown
 END
 
-killall synergy synergyc
+~/.scripts/sys_setkb.sh
